# Reduce contention in the work stealing pool for single fiber workloads