@djspiewak 2020-11-18
That's a pretty fantastic comparison chart; how did you generate that?