@vasilmkd 2020-11-18
> That's a pretty fantastic comparison chart; how did you generate that?

I used some VS Code multiple selection and copied the results into Numbers, then played around for a bit. 😄